genericarm64: clean up kernel modules and firmware
Don't install _all_ of the firmware, as that's huge (almost 1GB). Instead
install a few pieces of firmware for common hardware.

Also use the same list of packages to populate the initramfs, so there's
no need to manually sync the package lists (as initramfs doesn't install
the MACHINE_EXTRA_RRECOMMENDS automatically).

Signed-off-by: Ross Burton <ross.burton@arm.com>